[#5] Added JavaDoc

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/dto/SectorDto.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public abstract class SectorDto {
 
     private Long id;
-    private double price;
+    private Float price;
     private final SectorType type;
 
     public SectorDto() {
